fix: support localized custom route
